Iteration 2
In the second iteration of the code, moveRandomly() was added. The movement of the pattern is mainly achieved by `moveRandomly()`. The `moveRandomly()` method applies a random `(x, y)` coordinate offset when the music is playing, making the pattern produce a slight shaking effect. This random movement makes the pattern look like it is "jumping" with the music.