Iteration 3
In the third iteration, I mainly modified the moveRandomly() part. When the music plays, the pattern moves and bounces in random directions; and the boundary check ensures that the pattern moves within the canvas and bounces when it hits the boundary. By the way, I forgot to mention in the second iteration that when you pause in the second iteration, the text "play music to start" and the pattern will disappear.